Split module.cpp
